Refactor ExerciseCard to use context
