+++ src/components/ExerciseCard.module.css
@@ -8,6 +8,10 @@
   box-shadow: 0 2px 6px rgba(0,0,0,.08);
 }
 
+.active {
+  outline: 3px solid var(--color-primary);
+}
+
 .skeleton {
   width: 100%;
   padding-top: 100%;
