+++ src/components/ExerciseCard.tsx
@@ -1,21 +1,17 @@
-export interface Exercise {
-  id: string;
-  nombre: string;
-  zonaPrincipal: string;
-}
-
-export type Props = {
-  exercise: Exercise;
-  onSelect: (e: Exercise) => void;
-};
+import { Exercise, useExercises } from "../context/ExerciseContext";
 
 import styles from "./ExerciseCard.module.css";
 import { useState } from "react";
 
-export default function ExerciseCard({ exercise, onSelect }: Props) {
+export default function ExerciseCard({ exercise }: { exercise: Exercise }) {
+  const { toggle, selected } = useExercises();
   const [loaded, setLoaded] = useState(false);
+  const isSel = selected.some(e => e.id === exercise.id);
   return (
-    <button onClick={() => onSelect(exercise)} className={styles.card}>
+    <button
+      onClick={() => toggle(exercise.id)}
+      className={`${styles.card} ${isSel ? styles.active : ""}`}
+    >
       <div style={{ position: "relative" }}>
         {!loaded && <div className={styles.skeleton} />}
         <img
